R6 class for a plant protection product with at least one active ingredient

Contains basic information about the active ingredients in the product

Format

An R6Class generator object.

Public fields

name

The name of the product

ais

A list of active ingredients

concentrations

The concentration of the ais

concentration_units

Defaults to g/L

density

The density of the product

density_units

Defaults to g/L Creates a new instance of this R6 class.

...

Identifiers of the active ingredients

concentrations

Concentrations of the active ingredients

concentration_units

Defaults to g/L

density

The density

density_units

Defaults to g/L Printing method

Active bindings

...

Identifiers of the active ingredients

Methods

Public methods

ppp$new()

ppp$print()

ppp$clone()

Method new()

Usage

ppp$new(
  name,
  ...,
  concentrations,
  concentration_units = "g/L",
  density = 1000,
  density_units = "g/L"
)

Method print()

Usage

ppp$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

ppp$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.