An R6 class for pesticidal active ingredients and associated data

The class is initialised with an identifier which is generally an ISO common name. Additional chemical information is retrieved from the internet if available.

pai

Format

An R6Class generator object

Fields

iso

ISO common name according to ISO 1750 as retreived from www.alanwood.net/pesticides

alanwood

List of information retreived from www.alanwood.net/pesticides

Examples

atr <- pai$new("atrazine")

#> alanwood.net:

#> Querying atrazine.html

#> PubChem:

#> https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/inchikey/cids/JSON

#> https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/cid/property/MolecularFormula,MolecularWeight,CanonicalSMILES,IsomericSMILES,InChI,InChIKey,IUPACName,XLogP,TPSA,Complexity,Charge,HBondDonorCount,HBondAcceptorCount/JSON

#> https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/cid/synonyms/JSON

#> Trying to get chemical information from RDKit using PubChem_Canonical SMILES
#> CCNC1=NC(=NC(=N1)Cl)NC(C)C

#> Did not find chyaml file ./atrazine.yaml

print(atr)

#> <pai> with ISO common name $iso atrazine 
#> <chent>
#> Identifier $identifier atrazine 
#> InChI Key $inchikey MXWJVTOOROXGIU-UHFFFAOYSA-N 
#> SMILES string $smiles:
#>            PubChem_Canonical 
#> "CCNC1=NC(=NC(=N1)Cl)NC(C)C" 
#> Molecular weight $mw: 215.7 
#> PubChem synonyms (up to 10):
#>  [1] "2256"         "atrazine"     "1912-24-9"    "Gesaprim"     "Chromozin"   
#>  [6] "Oleogesaprim" "Aktikon"      "Argezin"      "Atazinax"     "Atranex"